+++ shared/locales/en-US.js
@@ -4,16 +4,16 @@ var version = "Shows azk version";
 var verbose = "Sets the level of detail";
 var quiet   = "Never prompt";
 //var systems_options = "Targets systems of action";
-var rebuild = "Force the rebuild or pull image and reprovision before starting an instance";
+var rebuild = "Force the rebuild, or pull image and reprovision before starting an instance";
 var reprovision = "Force the provisioning actions before starting an instance";
 
 // jscs:disable maximumLineLength
 module.exports = {
   errors: {
-    not_vm_start: "Unable to install and configure virtual machine",
-    not_connect_docker: "Could not initialize balancer because Docker was not available",
-    agent_not_running: "Azk agent is required but is not running (try `azk agent status`)",
-    agent_start: "Azk agent start error: %(error)s",
+    no_vm_started: "Unable to install and configure virtual machine",
+    connect_docker_unavailable: "Could not initialize balancer because docker was not available",
+    agent_not_running: "azk agent is required but is not running (try `azk agent status`)",
+    agent_start: "azk agent start error: %(error)s",
     not_been_implemented: "This feature: `%(feature)s` has not been implemented yet",
     system_not_found: "System `%(system)s` not found in `%(manifest)s`",
     manifest_required: "Manifest is required, but was not found in `%(cwd)s`",
@@ -21,20 +21,26 @@ module.exports = {
     required_option_error: "Option %(option)s is required",
     system_depend_error: "System `%(system)s` depends on the system `%(depend)s`",
     system_run_error: "Run system `%(system)s` return: (%(exitCode)d), for command: %(command)s:\n%(log)s",
-    system_not_scalable: "System `%(system)s is not scalable only one instance is allowed.",
-    image_not_available: "System `%(system)s` requires image `%(image)s` which is not avaible",
+    system_not_scalable: "System `%(system)s` is not scalable only one instance is allowed.",
+    image_not_available: "System `%(system)s` requires image `%(image)s` which is not available",
     run_command_error: "Run `%(command)s` in system `%(system)s` error:\n`%(output)s`",
     provision_pull_error: "Error downloading/pulling docker image `%(image)s`, message: %(msg)s.",
     invalid_option_error: "Invalid argument option: %(option)s",
     invalid_value_error: "Invalid value: %(value)s in option %(option)s",
-    image_not_exist: "Image from '%(image)s' not found",
-    provision_not_found: "Not found '%(image)s' image",
+    image_not_exist: "Image from '%(image)s' was not found",
+    provision_not_found: "Could not find '%(image)s' image",
     os_not_supported: "System not supported (see http://azk.io)",
+    run_timeout_error: [
+      "[timeout] `azk` has timed out on `%(system)s` system.",
+      "[timeout] Failure to reach port `%(port)s` from `%(hostname)s` after %(retry)s attempt[s] of %(timeout)s milliseconds.",
+      "[timeout] Make sure the start command binds `port` to the `0.0.0.0` interface, not only to the `localhost` interface.",
+      "[timeout] You might want to edit your `Azkfile.js` in order to increase the maximum timeout.",
+    ].join("\n"),
 
     docker_build_error: {
       server_error: "Internal error in build `%(dockerfile)s`: %(error)",
       not_found   : "Can't find `%(from)s` image to build `%(dockerfile)s`",
-      can_find_dockerfile: "Can't find %(dockerfile)s file",
+      can_find_dockerfile: "Can't find `%(dockerfile)s` file",
       can_find_add_file_in_dockerfile: "Can't find `%(source)s` file to ADD in `%(dockerfile)s`",
     },
 
@@ -43,23 +49,23 @@ module.exports = {
         upgrade: ["\nYou are using `v%(current_version)s` version.",
                   "`v%(new_version)s` version is available.",
                   "Please, access http://azk.io to upgrade\n"].join("\n"),
-        mv_resolver: "Upgrading domains error, was not possible to move files",
+        mv_resolver: "Upgrading domains error, moving files was not possible",
       },
       darwin: {
-        VBoxManage : 'VirtualBox not installed. Install before continue.',
+        VBoxManage : 'VirtualBox not installed. Install before continuing.',
         unfsd: 'unfs3 not installed. Reinstall `azk` or `brew install unfs3` before continuing.',
-        network: 'networking error',
+        network: 'Networking error',
       },
       linux: {
         port_error: [
           "The %(port)s port configured for the `%(service)s` service is not available.",
-          "Try configure a different port setting %(env)s environment variable.",
+          "Try configuring a different port setting %(env)s environment variable.",
           "Check http://docs.azk.io for more information.",
         ].join('\n'),
         docker_access: [
           "Could not connect to the docker service.",
           "Check if docker service is running.",
-          "And check if you have write permission to socket: '%(socket)s'",
+          "Also check if you have write permission to socket: '%(socket)s'",
         ].join('\n'),
       }
     }
@@ -70,7 +76,7 @@ module.exports = {
       running: "Agent is running...",
       not_running: "Agent is not running (try: `azk agent start`).",
       starting: "Agent is being started...",
-      already: "Agent is already running.",
+      already_running: "Agent is already running.",
       started: "Agent has been successfully started.",
       stopping: "Agent is being stopped...",
       stopped: "Agent has been successfully stopped.",
@@ -85,22 +91,22 @@ module.exports = {
       starting    : "Starting virtual machine...",
       started     : "Virtual machine has been successfully started.",
       // TODO: stopping
-      stoping     : "Stopping virtual machine...",
+      stopping     : "Stopping virtual machine...",
       // TODO: stopped
-      stoped      : "Virtual machine has been successfully stopped.",
+      stopped      : "Virtual machine has been successfully stopped.",
       removing    : "Removing virtual machine...",
       removed     : "Virtual machine has been successfully removed.",
       // TODO: waiting
       wait        : "Waiting for initialization of virtual machine...",
       initialized : "Virtual machine is ready to use.",
-      progress    : "Trying connect to vm (%(uri)s) (%(attempts)d/%(max)d)...",
-      upkey       : "Upload the ssh key to vm...",
+      progress    : "Trying to connect to vm (%(uri)s) (%(attempts)d/%(max)d)...",
+      upkey       : "Uploading the ssh key to vm...",
       error       : "Error in vm process: %(data)s",
-      docker_keys : "Downloading required keys to connect with docker",
+      docker_keys : "Downloading required keys to connect to docker",
     },
 
     socat: {
-      progress : "trying to connect to docker (%(attempts)d/%(max)d)...",
+      progress : "Trying to connect to docker (%(attempts)d/%(max)d)...",
     },
 
     'balancer-redirect_connect': {
@@ -111,60 +117,74 @@ module.exports = {
       starting_memcached: "Starting memcached service...",
       started_memcached : "Memcached service started.",
       stopping_memcached: "Stopping memcached service...",
-      stoped_memcached  : "Memcached service was stoped.",
+      stopped_memcached  : "Memcached service was stopped.",
       exited_memcached  : "Memcached service was `exited`.",
 
       starting_hipache  : "Starting http balancer service...",
       started_hipache   : "Http balancer service started.",
       stopping_hipache  : "Stopping http balancer service...",
-      stoped_hipache    : "Http balancer service was stoped.",
+      stopped_hipache    : "Http balancer service was stopped.",
       exited_hipache    : "Http balancer service was `exited`.",
 
       'starting_balancer-redirect': "Starting azk balancer redirect service...",
       'started_balancer-redirect' : "Balancer redirect started.",
       'stopping_balancer-redirect': "Stopping balancer redirect...",
-      'stoped_balancer-redirect'  : "Balancer redurect was stoped.",
+      'stopped_balancer-redirect'  : "Balancer redirect was stopped.",
 
       starting_dns      : "Starting azk dns service...",
       started_dns       : "Dns service started.",
       stopping_dns      : "Stopping dns service...",
-      stoped_dns        : "Dns service was stoped.",
-      progress          : "Trying connect to docker (%(uri)s) (%(attempts)d/%(max)d)...",
+      stopped_dns        : "Dns service was stopped.",
+      progress          : "Trying to connect to docker (%(uri)s) (%(attempts)d/%(max)d)...",
     },
 
     unsfd: {
-      starting : "starting unsfd...",
+      starting : "Starting unsfd...",
       started  : "unsfd started in %(port)s port with file config: %(file)s.",
-      stopping : "stopping unsfd...",
+      stopping : "Stopping unsfd...",
       // TODO: stopped
-      stoped   : "unsfd has been successfully stopped.",
-      mounting : "mounting the unsfd shared folder in virtual machine...",
+      stopped   : "unsfd has been successfully stopped.",
+      mounting : "Mounting the unsfd shared folder in virtual machine...",
       mounted  : "unsfd shared folder has been successfully mounted.",
     },
   },
 
   generator: {
-    found: "`%(__type)s` system was detected at '%(dir)s' as '%(systemName)s'",
+    found              : [
+      "",
+      "[%(systemName)s] `A %(__type)s` system was detected at '%(dir)s'.",
+      "[%(systemName)s] The image suggested was `%(image)s`.",
+    ].join("\n"),
+    foundWithoutVersion: [
+      "",
+      "[%(systemName)s] A `%(__type)s` system was detected at '%(dir)s'.",
+      "[%(systemName)s] The image suggested was `%(image)s`.",
+      "[%(systemName)s] ! It was not possible to detect the `%(__type)s` specific version, so the standard version was suggested instead.",
+      "[%(systemName)s] ! To change the image version you must edit the `Azkfile.js` file.",
+      "[%(systemName)s] ! For more information see the documentation at http://docs.azk.io/en/images/index.html.",
+    ].join("\n")
   },
 
   manifest: {
-    not_found: "no such '%s' in current project",
-    circular_depends: "Circular dependency between %(system1)s and %(system2)s",
-    image_required: "Not image set for the `%(system)s' system",
-    can_find_dockerfile: "Can't find `%(dockerfile)s` file to build a image for `%(system)s` system",
-    system_name_invalid: "The system name `%(system)s` is not valid.",
-    provider_invalid: "The provider not found: `%(wrongProvider)s`.",
-    depends_not_declared: "The `%(system)s` system depends on the `%(depend)s` system, which was not stated.",
-    balancer_depreciation: "The `balancer` option used in the `%(system)s` is deprecated, use `http` and `scalable` to replace",
-    invalid_default: "Unable to set the system `%(system)s` as a default because it was not declared",
-    mount_and_persistent_depreciation: [
+    balancer_deprecated   : "The `balancer` option used in the `%(system)s` is deprecated, use `http` and `scalable` instead",
+    cannot_extends_itself : "The system `%(system)s` cannot extend itself",
+    cannot_find_dockerfile: "Can't find `%(dockerfile)s` file to build an image for `%(system)s` system",
+    circular_dependency   : "Circular dependency between %(system1)s and %(system2)s",
+    depends_not_declared  : "The `%(system)s` system depends on the `%(depend)s` system, which was not declared.",
+    extends_system_invalid: "The system `%(system_source)s` for extending system `%(system_to_extend)s` cannot be found",
+    image_required        : "No image set for the `%(system)s' system",
+    invalid_default       : "Unable to set the system `%(system)s` as a default because it was not declared",
+    mount_and_persistent_deprecated: [
       "The `%(option)s` option used in system `%(system)s` is no longer supported.",
       "You must change the %(manifest)s to use `mounts`",
       "Check http://git.io/29JW0w for further information",
     ].join("\n"),
+    not_found             : "No such '%s' in current project",
+    provider_invalid      : "The provider was not found: `%(wrongProvider)s`.",
+    system_name_invalid   : "The system name `%(system)s` is not valid.",
     validate: {
       deprecated : "The `%(option)s` used in `%(system)s` is deprecated, check the documentation for `%(new_option)s`",
-      not_systems: "No system has been set yet, check the documentation",
+      no_system_set: "No system has been set yet, check the documentation",
     }
   },
 
@@ -178,23 +198,23 @@ module.exports = {
     loading_checking: "Loading settings and checking dependencies.",
     ip_question: "Enter the vm ip",
     ip_invalid: "`%(ip)s`".yellow + " is an invalid v4 ip, try again.",
-    ip_of_range: "`%(ip)s`".yellow + " is an invalid ip range, try again.",
+    ip_invalid_range: "`%(ip)s`".yellow + " is an invalid ip range, try again.",
     adding_ip: "Adding %(ip)s to %(file)s ...",
-    generating_key: "Generating public/private rsa key pair to connect vm.\n",
+    generating_key: "Generating public/private rsa key pair to connect to vm.\n",
     vm_ip_msg: ([
       "",
       "In order to give `azk` access to `azk agent`,",
       "it is necessary to define an IP address to the virtual machine.",
       "This IP address will be used to establish a private network",
       "between the physical machine running `azk` and the virtual",
-      "machine where `azk agent` is in execution.",
+      "machine where `azk agent` is running.",
       "",
     ]).join('\n'),
     check_version: 'Checking version...',
-    check_version_error: 'checking version: [ %(error_message)s ]!',
+    check_version_error: 'Checking version: [ %(error_message)s ]!',
     clean_containers: "Clearing %(count)d lost containers",
     migrations: {
-      alert: "Azk updated, checking update procedures...",
+      alert: "azk updated, checking update procedures...",
       changing_domain: "Changing domain upgrading, (issue: #255)",
       moving_resolver: "Moving %(origin)s to %(target)s ...",
       renaming_vm: "Renaming VirtualBox machine %(old_name)s to %(new_name)s",
@@ -204,7 +224,7 @@ module.exports = {
   commands: {
     not_found: "Command '%s' not found",
     azk: {
-      description: "Azk short help",
+      description: "azk short help",
       options: {
         command: {
           name: "command".yellow
@@ -227,7 +247,7 @@ module.exports = {
     agent: {
       description: "Controls azk agent",
       start_fail: "Agent start fail: %s",
-      start_before: "The agent is not running, would you like to start it now?",
+      start_before: "The agent is not running, would you like to start it?",
       options: {
         verbose: verbose,
         quiet: quiet,
@@ -235,8 +255,8 @@ module.exports = {
           name: "actions".magenta,
           options: {
             start: "Start azk agent",
-            status: "Show the azk agent status",
-            stop: "Stop the azk agent running in the background",
+            status: "Shows azk agent status",
+            stop: "Stops azk agent running in the background",
           }
         },
         'reload-vm': "Reloads the virtual machine settings",
@@ -244,25 +264,25 @@ module.exports = {
       }
     },
     configs: {
-      description: "Shows the azk configs values",
+      description: "Shows azk configs values",
     },
     docker: {
       options: {
         verbose: verbose,
         quiet: quiet,
       },
-      description: "Alias for calling the docker in 'azk' scope configurations",
+      description: "Alias for calling docker in 'azk' scope configuration",
     },
     doctor: {
-      description: "Shows an analysis of the health of `azk`",
+      description: "Shows an analysis of `azk`'s health",
       options: {
-        logo: "Shows the `azk` logo before show health informations",
+        logo: "Shows the `azk` logo before showing health information",
         verbose: verbose,
         quiet: quiet,
       },
     },
     info: {
-      description: "Shows systems informactions for the current `Azkfile.js`",
+      description: "Shows systems information for the current `Azkfile.js`",
       options: {
         verbose: verbose,
         quiet: quiet,
@@ -280,7 +300,7 @@ module.exports = {
       },
     },
     shell: {
-      description: "Initializes a shell with instance context or runs a arbitrary command",
+      description: "Initializes a shell with instance context, or runs a specified command",
       invalid_mount: "Invalid mount parameter: `%(value)s`, use `point:[path:|persitent:]origin`",
       invalid_env: "Invalid env variable: `%(value)s`, use `VARIABLE=VALUE`",
       options: {
@@ -289,19 +309,19 @@ module.exports = {
         system  : "A system context to execute a shell or command",
         remove  : "Removes shell instances after exit shell or command",
         command : "Runs a specific command",
-        shell   : "The path to shell binary",
+        shell   : "The path to the shell binary",
         silent  : "Prevents any log message about the command execution. It's useful when using the `-c` option and the output is used as input to another command using the pipe `|` operator.",
         verbose : verbose,
         quiet   : quiet,
-        mount   : "Points for additional mounting (ex:./origin:/azk/target)",
+        mount   : "Points to additional mounting (ex:./origin:/azk/target)",
         cwd     : "Default directory",
         image   : "Defines the image in which the command will be executed",
         env     : "Additional environment variables",
       },
       ended: {
         removed: "finished, because the container was removed",
-        docker_end: "finished, because the docker was finalized",
-        docker_notfound: "finished, because docker not found",
+        docker_end: "finished, because docker was finalized",
+        docker_not_found: "finished, because docker was not found",
       },
       examples: [
         '$ azk shell --shell /bin/bash',
@@ -311,7 +331,7 @@ module.exports = {
       ]
     },
     help: {
-      description: "Shows help about the specific command",
+      description: "Shows help about a specific command",
       usage: 'Usage:'.blue + ' $ %s',
       options: "options:".green,
       examples: "examples:".yellow,
@@ -321,29 +341,29 @@ module.exports = {
         pulling: 'Pulling repository %s...',
         bar_progress: '  :title [:bar] :percent :progress',
         bar_status: '  :title :msg',
-        pull_ended  : "✓".blue + " complete downloading `" + "%(image)s".yellow + "`",
+        pull_ended  : "✓".blue + " completed download of `" + "%(image)s".yellow + "`",
       }
     },
     init: {
       description: "Initializes a project by adding the file Azkfile.js",
-      already: "'%s' already exists (try: `--force`)",
+      already_exists: "'%s' already exists (try: `--force`)",
       generated: "'%s' generated",
-      github: "\nTip:\n  Adds the `.azk` in .gitignore\n  echo '.azk' >> .gitignore \n",
-      not_found: "Not found a system(s), generating with example system.",
+      github: "\nTip:\n  Adds the `.azk` to .gitignore\n  echo '.azk' >> .gitignore \n",
+      not_found: "System(s) not found, generating with example system.",
       options: {
-        force: "Forces the overlay file Azkfile.js",
+        force: "Forces overwrite of Azkfile.js",
       }
     },
     start: {
       description: "Starts an instance of the system(s)",
-      already: "System `%(name)s` already started",
-      fail: "Due to the above error azk will stop all instances already begun.\n",
+      already_started: "System `%(name)s` already started",
+      fail: "Due to the above error azk will stop all instances already running.\n",
       options: {
         verbose: verbose,
         rebuild: rebuild,
         quiet   : quiet,
         reprovision: reprovision,
-        open: "Open a url of default system in the preferred application",
+        open: "Open a url of a default system in the preferred application",
       },
       option_errors: {
         open: {
@@ -358,13 +378,13 @@ module.exports = {
       options: {
         verbose : verbose,
         quiet   : quiet,
-        remove: "Removes the instances before stop",
+        remove: "Removes the instances before stopping",
       }
     },
     scale: {
       instances   : "from " + "%(from)d".red + " to " + "%(to)d".green + " instances",
       description : "Scales (up or down) an instance of the system(s)",
-      wait_port   : "◴".magenta + " waiting start `" + "%(system)s".blue  + "` system, try connect port %(name)s/%(protocol)s...",
+      wait_port   : "◴".magenta + " waiting for `" + "%(system)s".blue  + "` system to start, trying connection to port %(name)s/%(protocol)s...",
       check_image : "✓".cyan    + " checking `"      + "%(image)s".yellow + "` image...",
       pull_image  : "⇲".blue    + " downloading `"   + "%(image)s".yellow + "` image...",
       build_image : "⇲".blue    + " building `"      + "%(image)s".yellow + "` image...",
@@ -374,24 +394,24 @@ module.exports = {
       scaling_up  : "↑".green   + " scaling `"       + "%(system)s".blue  + "` system %(instances)s...",
       scaling_down: "↓".red     + " scaling `"       + "%(system)s".blue  + "` system %(instances)s...",
       options: {
-        remove: "Removes the instances before stop",
+        remove: "Removes the instances before stopping",
         verbose: verbose,
         quiet: quiet,
       }
     },
     restart: {
-      description: "Stops all system and starts again",
+      description: "Stops all systems and starts them again",
       options: {
         verbose: verbose,
         rebuild: rebuild,
         quiet   : quiet,
         reprovision: reprovision,
-        open: "Open a url of default system in the preferred application",
+        open: "Open a url of the default system in the preferred application",
       }
     },
     reload: {
-      description: "Stops all system, re-provisions and starts again",
-      deprecation: "`reload` this deprecated, use restart",
+      description: "Stops all systems, run provision commands and starts them again",
+      deprecation: "`reload` is deprecated, use restart",
       options: {
         verbose: verbose,
         rebuild: rebuild,
@@ -418,14 +438,14 @@ module.exports = {
     },
     vm: {
       description  : "Controls a virtual machine.",
-      already      : "virtual machine already installed.",
-      not_installed: "virtual machine is not installed, try `azk vm install`.",
-      running      : "virtual machine running.",
-      already_running : "virtual machine already running.",
+      already_installed  : "Virtual machine already installed.",
+      not_installed: "Virtual machine is not installed, try `azk vm install`.",
+      running      : "Virtual machine running.",
+      already_running : "Virtual machine already running.",
       // TODO not_running
-      not_runnig   : "virtual machine is not running, try `azk vm start`.",
+      not_running  : "Virtual machine is not running, try `azk vm start`.",
       error        : "vm error: %(error)s.",
-      not_requires : "this system not requires virtual machine, to try force this behavior set `AZK_USE_VM=true`",
+      not_required : "This system does not require a virtual machine, to try to force this behavior set `AZK_USE_VM=true`",
       options: {
         verbose : verbose,
         quiet   : quiet,
@@ -445,15 +465,15 @@ module.exports = {
   },
 
   docker: {
-    connect: "connecting to docker: %s..."
+    connect: "Connecting to docker: %s..."
   },
 
   proxy: {
-    adding_backend: "adding backend:%(backend)s to hostname:%(hostname)s...",
-    removing_backend: "removing backend:%(backend)s to hostname:%(hostname)s...",
-    request: "proxy request %s => %s",
-    started: "proxy started in %s port",
-    not_configured: "host '%s' not configured in proxy"
+    adding_backend: "Adding backend:%(backend)s to hostname:%(hostname)s...",
+    removing_backend: "Removing backend:%(backend)s to hostname:%(hostname)s...",
+    request: "Proxy request %s => %s",
+    started: "Proxy started in %s port",
+    not_configured: "Host '%s' not configured in proxy"
   },
 
   test: {