+++ src/generator/rules/python.js
@@ -1,4 +1,4 @@
-import { _ } from 'azk';
+import { _, isBlank } from 'azk';
 import { BaseRule } from 'azk/generator/rules';
 var semver = require('semver');
 
@@ -42,7 +42,7 @@ export class Rule extends BaseRule {
     evidence.version = pythonVersion;
 
     // cant find version, will use latest
-    if (pythonVersion === null) {
+    if (isBlank(pythonVersion)) {
       return evidence;
     }